refactor(locale): rename ne to en_NP
