+++ src/locale/en_NP.ts
@@ -1,0 +1,25 @@
+/*
+ * This file is automatically generated.
+ * Run 'pnpm run generate:locales' to update.
+ */
+
+import { Faker } from '../faker';
+import base from '../locales/base';
+import en from '../locales/en';
+import en_NP from '../locales/en_NP';
+
+/**
+ * The faker instance for the `en_NP` locale.
+ *
+ * - Language: English (Nepal)
+ * - Endonym: English (Nepal)
+ *
+ * This instance uses the following locales internally (in descending precedence):
+ *
+ * - `en_NP`
+ * - `en`
+ * - `base`
+ */
+export const faker = new Faker({
+  locale: [en_NP, en, base],
+});
