+++ src/locale/ne.ts
@@ -1,25 +1,26 @@
-/*
- * This file is automatically generated.
- * Run 'pnpm run generate:locales' to update.
- */
-
-import { Faker } from '../faker';
-import base from '../locales/base';
-import en from '../locales/en';
-import ne from '../locales/ne';
+import { deprecated } from '../internal/deprecated';
+import { faker as en_NP } from './en_NP';
 
 /**
- * The faker instance for the `ne` locale.
+ * The faker instance for the `en_NP` locale.
  *
- * - Language: Nepali
- * - Endonym: नेपाली
+ * - Language: English (Nepal)
+ * - Endonym: English (Nepal)
  *
  * This instance uses the following locales internally (in descending precedence):
  *
- * - `ne`
+ * - `en_NP`
  * - `en`
  * - `base`
+ *
+ * @deprecated Use `fakerEN_NP` instead. This export will be removed in version 11.0.
  */
-export const faker = new Faker({
-  locale: [ne, en, base],
-});
+export const faker = (() => {
+  deprecated({
+    deprecated: "import { faker } from '@faker-js/faker/locale/ne'",
+    proposed: "import { faker } from '@faker-js/faker/locale/en_NP'",
+    since: '10.5',
+    until: '11.0',
+  });
+  return en_NP;
+})();
