+++ src/locales/en_NP/company/index.ts
