+++ src/locales/en_NP/company/legal_entity_type.ts
