+++ src/locales/en_NP/internet/domain_suffix.ts
