+++ src/locales/en_NP/internet/index.ts
