+++ src/locales/en_NP/location/city_name.ts
