+++ src/locales/en_NP/location/city_pattern.ts
