+++ src/locales/en_NP/location/index.ts
