+++ src/locales/en_NP/location/state.ts
