+++ src/locales/en_NP/location/street_pattern.ts
