+++ src/locales/en_NP/metadata.ts
@@ -1,0 +1,13 @@
+import type { PreBuiltMetadataDefinitionForCountry } from '../../definitions/metadata';
+
+const metadata: PreBuiltMetadataDefinitionForCountry = {
+  title: 'English (Nepal)',
+  code: 'en_NP',
+  country: 'NP',
+  language: 'en',
+  endonym: 'English (Nepal)',
+  dir: 'ltr',
+  script: 'Latn',
+};
+
+export default metadata;
