+++ src/locales/en_NP/person/index.ts
