+++ src/locales/en_NP/person/last_name_pattern.ts
