+++ src/locales/en_NP/phone_number/format/human.ts
