+++ src/locales/en_NP/phone_number/format/national.ts
