+++ src/locales/en_NP/phone_number/index.ts
