--- src/locales/ne/metadata.ts
@@ -1,12 +1,0 @@
-import type { PreBuiltMetadataDefinition } from '../../definitions/metadata';
-
-const metadata: PreBuiltMetadataDefinition = {
-  title: 'Nepali',
-  code: 'ne',
-  language: 'ne',
-  endonym: 'नेपाली',
-  dir: 'ltr',
-  script: 'Deva',
-};
-
-export default metadata;
